a) "Our cooperative currency, Ithaca Dollars (i$), work…" (edited)
b) Please make sure the How It Works marquee cycles back around after completion. It stopped on the last slide for me

c) Can we increase the height of each band in the Getting Started section? The text feels pinched.

f) The “Network” should be “The Circuit” so as to keep consistent in our languaging.

h) the “buy i$” button in the header goes to “apply”. please fix to send to our i$ sales page.

i) I’d like to have the “Buy i$” button be the dark green instead of the same light green as “apply"

j) please delete the exclamation point from the “Join the Circuit!” buttons

l) delete the Ithacash graphic on the left side of the Say Hello section, move the “A MEMBER OF” text closer to the top of that section, and add in the following icons + links: