Per Scott on Slack:
d) Specifically for 3c, 3d, 4, 5, + 6, can you please play around with which words are on which lines to make the paragraph more … together? i.e. even line lengths, no single words on the last line, etc.

e) For the email “Share” icon, it seems to make more sense to equip them with an email they can send to a friend than an email that’s sent to us. Can we hammer out what this would look like?

k) please add some room between the bulleted list of ways each member gets i$ and the Join the Circuit button

m) we need to add a checkbox / dropdown on the apply page so that people can tell us what kind of account they’re applying for.