Don't require email confirmation to move forward with signup
Note: The following implementation is tacked on to the original flow. A more robust implementation would take the non-mandatory email confirmation into account from the ground up. That more robust implementation would be an opportunity for a more robust API.

We now do not require that the user confirm their email before being able to complete the signup flow. Now, the user is directed directly to the Account Selection step immediately after entering their email. They still receive an email asking to confirm their email address.

We've created an api endpoint for email confirmations that is sent in the email confirmation email. On successful confirmation, this will redirect them to thetheir respect create_account page.